+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 
+- Slack reminder example
 - Type check and linting pre-commits
 
 ## [0.0.4] - 2023-06-09